change default firmware partition size to dual partition size
Change the size of the firmware partition from 16000kB to 7936kB.
This size is the same used by the bootloader. The rationale behind
this change is that if for any reason the boot process doesn't use
the correct partitions provided by the bootloader in mtdparts then
this size will be used and with a 16000kB size (only one big partiton)
default then the image will boot and jffs2 will corrupt the "stable"
partition.